+++ src/main/java/me/despical/commonsbox/number/NumberUtils.java
@@ -1,19 +1,19 @@
 /*
- * CommonsBox - Library box of common utilities.
+ * Commons Box - Box of common utilities.
  * Copyright (C) 2020 Despical
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
+ *  it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
 package me.despical.commonsbox.number;
@@ -27,42 +27,98 @@ public class NumberUtils {
 
 	private NumberUtils() {}
 
-	public static boolean isInteger(String s) {
+	/**
+	 * Checks if the String contains only unicode digits.
+	 * A decimal point is not a unicode digit and returns false.
+	 *
+	 * @param str the String to check, may be null
+	 * @return    true if only contains digits, and is non-null
+	 */
+	public static boolean isInteger(String str) {
+		if (str == null) {
+			return false;
+		}
+
+		int length = str.length();
+
+		if (length == 0) {
+			return false;
+		}
+
+		int i = 0;
+
+		if (str.charAt(0) == '-') {
+			if (length == 1) {
+				return false;
+			}
+
+			i = 1;
+		}
+
+		for (; i < length; i++) {
+			char c = str.charAt(i);
+
+			if (c <= '/' || c >= ':') {
+				return false;
+			}
+		}
+
+		return true;
+	}
+
+	/**
+	 * Checks if the String is primitive type double or not.
+	 *
+	 * @param str the String to check, may be null
+	 * @return    true if double, and is non-null
+	 */
+	public static boolean isDouble(String str) {
 		try {
-			Integer.parseInt(s);
+			Double.parseDouble(str);
 			return true;
-		} catch (NumberFormatException ex) {
+		} catch (NumberFormatException ignored) {
 			return false;
 		}
 	}
 
-	public static boolean isDouble(String s) {
+	/**
+     * Checks if the String is primitive type long or not.
+	 *
+	 * @param str the String to check, may be null
+	 * @return    true if long, and is non-null
+	 */
+	public static boolean isLong(String str) {
 		try {
-			Double.parseDouble(s);
+			Long.parseLong(str);
 			return true;
-		} catch (NumberFormatException ex) {
+		} catch (NumberFormatException ignored) {
 			return false;
 		}
 	}
 
-	public static boolean isLong(String s) {
+	/**
+	 * Checks if the String is primitive type short.
+	 *
+	 * @param str the String to check, may be null
+	 * @return    true if short, and is non-null
+	 */
+	public static boolean isShort(String str) {
 		try {
-			Long.parseLong(s);
+			Short.parseShort(str);
 			return true;
-		} catch (NumberFormatException ex) {
+		} catch (NumberFormatException ignored) {
 			return false;
 		}
 	}
 
-	public static boolean isShort(String s) {
-		try {
-			Short.parseShort(s);
-			return true;
-		} catch (NumberFormatException ex) {
-			return false;
-		}
-	}
-
+	/**
+	 * Checks if the given value is between given values.
+	 *
+	 * @param value the Integer to check is between values
+	 * @param min   the minimum value to check given Integer
+	 * @param max   the maximum value to check given Integer
+	 * @return      true if value is between min and max values
+	 */
 	public static boolean isBetween(int value, int min, int max) {
 		return value >= min && value <= max;
 	}
@@ -78,6 +134,6 @@ public class NumberUtils {
 	 */
 	public static int serializeInt(Integer i) {
 		if (i == 0) return 9;
-		return (i % 9) == 0 ? i : (i + 9 - 1) / 9 * 9;
+		return i % 9 == 0 ? i : (i + 9 - 1) / 9 * 9;
 	}
 }